Skyscope-AI-ByBit-Affiliate-Marketing-Agent/ai_marketing_agent/requirements.txt
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
opencv-python
PyYAML
pyzbar
requests
rich
